Create storage class
